backend/requirements.txt
Django==5.0.4
django-environ==0.11.2
pytest-django==4.8.0
djangorestframework==3.15.1
django-cors-headers==4.3.1
